stop if no users found in remove

+++ src/main/java/aggregation/ProcessingManager.java
@@ -60,6 +60,9 @@ public class ProcessingManager {
 	
 	public void removeMarkedData() {
 		Collection<User> users = remover.findUsersAffectedByRemoval();
+		if (users.isEmpty()) {
+			return;
+		}
 		Collection<String> zipsToReimport = remover.findAllZipsForUsers(users);
 		remover.deleteZipsToDelete();
 		remover.deleteAllDataForUsers(users);
